feat: Update SSelectBox component to support custom options
The SSelectBox component has been updated to support custom options. This allows for more flexibility in the options that can be displayed in the select box. The `useCustomOptions` prop has been added to the component, which when set to `true`, will use the custom options provided instead of the default options.

This change also includes updates to the InterventionTeams component, where the select box options are now populated with user profiles. The options are now an array of arrays, with each inner array containing the display text and the corresponding value. The `name` prop has been replaced with `id` to reflect this change.

This commit addresses the need for custom options in the SSelectBox component and improves the functionality of the InterventionTeams component.